Move `CoreDataFeedImageStore` API's into existing `CoreDataFeedStore`
